Busy Cursor and a Qt Workaround
(1) The cursor is set to busy after one second of unfinished activity when a huge number of files or a huge file is opened/closed.

(2) Qt doesn't disbale the shortcuts of the main window when it has a modal dialog. A workaround is added for this bug because some shortcuts can cause a crash.

(3) Dialogs and some shortcuts are deactivated while files are being loaded.